Split MFSDP context from module mixin
